src: shuffle around 'pipe'
